cmake imrovements

+++ threading/src/buffers.cpp
@@ -15,7 +15,7 @@ namespace Nlib
 		}
 		bool CyclicIndex::operator>(const CyclicIndex& t) const noexcept
 		{
-			constexpr uint16_t uint8max = ~static_cast<uint8_t>(0), uint8maxWrapped = ~static_cast<uint16_t>(0) - uint8max;
+			constexpr uint16_t uint8max = ~static_cast<uint8_t>(0), uint8maxWrapped = (~static_cast<uint16_t>(0)) - static_cast<uint16_t>(uint8max);
 			return (index < uint8max && t.index > uint8maxWrapped) ? true : index > t.index;
 		}
 		bool CyclicIndex::operator>=(const CyclicIndex& t) const noexcept
